fix: fix type Salary amount field and add type AverageSalary (#682)

+++ src/schema/experience.js
@@ -373,7 +373,7 @@ const Mutation = gql`
 
     input SalaryInput {
         type: SalaryType!
-        amount: Float!
+        amount: Int!
     }
 
     input SectionInput {
